server/requirements.txt
flask
sqlalchemy
readability-lxml
psycopg2
langdetect
